+++ astrid/core/pack_machinery/__init__.py
@@ -1,0 +1,8 @@
+"""Pack machinery — internal implementation modules for pack operations.
+
+These are the canonical homes for pack-operation machinery moved from
+``astrid/packs/`` during M1 Pack Layout Normalization (Plan v1.0).
+
+Public APIs are re-exported through thin compatibility shims in
+``astrid/packs/`` for backward compatibility until M2 clean-up.
+"""
