+++ astrid/core/pack_machinery/schemas/v1/executor.json
