+++ astrid/core/pack_machinery/schemas/v1/orchestrator.json
