+++ astrid/core/pack_machinery/schemas/v1/pack.json
