+++ astrid/packs/builtin/pack.yaml
@@ -17,3 +17,10 @@ agent:
   purpose: "Historical core namespace retained only for backward-compatible aliases (builtin.*). It ships no live components of its own."
   do_not_use_for: "Anything real — pick the pack that now owns the capability (editorial for cut/transcript work, foley for spatial audio, generation for image/video synthesis, fal for fal.ai)."
   normal_entrypoints: []
+metadata:
+  layout:
+    exceptions:
+      - path: agent_probe.py
+        class: dsl_orchestrator_shim
+        reason: "DSL @orchestrator definition with 16+ test references; relocation deferred to M2."
+        defer_to: M2
