--- astrid/packs/media/STAGE.md
@@ -1,11 +1,0 @@
-# Media
-
-## Purpose
-
-Lossless video clip extraction via ffmpeg stream copy. Use when you need to
-trim a segment from a video without re-encoding — fast and quality-preserving.
-
-## Components
-
-- **Executors**: `media.clip_extract` — extracts a clip given start time and duration.
-- **Orchestrators**: none.
