+++ astrid/packs/text_analysis/orchestrators/summarize/STAGE.md
@@ -1,0 +1,4 @@
+# Text Analysis Summarize
+
+Reads the pack's bundled `sample.txt` fixture and writes `content.json`,
+`summary.json`, and `verdict.json` into the orchestrator output directory.
