+++ astrid/packs/text_analysis/orchestrators/summarize/orchestrator.yaml
@@ -1,0 +1,35 @@
+{
+  "schema_version": 1,
+  "id": "text_analysis.summarize",
+  "name": "Text Analysis Summarize",
+  "kind": "built_in",
+  "version": "0.1.0",
+  "short_description": "Summarize the bundled sample text fixture into content, summary, and verdict JSON outputs.",
+  "description": "Read the text_analysis sample fixture, emit content and summary metadata, and write a one-line verdict artifact.",
+  "keywords": ["text", "summary", "analysis", "verdict"],
+  "outputs": [
+    {"name": "content", "type": "file", "path_template": "{out}/content.json"},
+    {"name": "summary", "type": "file", "path_template": "{out}/summary.json"},
+    {"name": "verdict", "type": "file", "path_template": "{out}/verdict.json"}
+  ],
+  "child_executors": [],
+  "child_orchestrators": [],
+  "runtime": {
+    "kind": "command",
+    "command": {
+      "argv": [
+        "{python_exec}",
+        "-m",
+        "astrid.packs.text_analysis.orchestrators.summarize.run",
+        "{orchestrator_args}"
+      ]
+    }
+  },
+  "cache": {"mode": "none"},
+  "metadata": {
+    "runtime_entrypoint": "main",
+    "runtime_file": "run.py",
+    "runtime_module": "astrid.packs.text_analysis.orchestrators.summarize.run",
+    "requires_output_path": true
+  }
+}
