+++ docs/megaplan/epics/pack-taxonomy/briefs/m2-alias-compatibility.md
@@ -61,7 +61,7 @@ machinery.
 
 - `astrid/core/alias_resolver.py`
 - executor/orchestrator registries
-- `astrid/packs/schemas/v1/pack.json`
+- `astrid/core/pack_machinery/schemas/v1/pack.json`
 - `astrid/packs/validate.py`
 - `astrid/packs/cli.py`
 - docs for aliases/deprecation
